feat: add Hygraphlix to settings and refactor import paths for consistency across the project
