+++ .vscode/settings.json
@@ -30,6 +30,7 @@
     "arya",
     "datatables",
     "Dorne",
+    "Hygraphlix",
     "Lançassolar",
     "martell",
     "Martell",
