+++ apps/oberyn/src/components/MovieCard.tsx
@@ -4,7 +4,7 @@ import { Card, CardHeader, CardBody } from '@nextui-org/react'
 import Image from 'next/image'
 import Link from 'next/link'
 
-import { MovieCardProps } from '@/src/types/types'
+import { MovieCardProps } from '@/types'
 
 const MovieCard: React.FC<MovieCardProps> = ({
   Poster,
